UI changes in places management
