+++ frontend-argonout/src/components/admin/routeEditorPanel/routeEditorPanel.css
@@ -1,0 +1,61 @@
+.AppContent {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    min-height: 100vh;
+}
+
+.right-top-container {
+    position: fixed; /* Ustawienie na stałe */
+    top: 0;          /* Przylega do górnej krawędzi */
+    right: 0;        /* Przylega do prawej krawędzi */
+    box-sizing: border-box; /* Uwzględnienie paddingu w szerokości */
+    z-index: 1000;   /* Opcjonalne: aby znajdował się nad innymi elementami */
+}
+
+.add-place-form {
+    font-size: 3vh;
+}
+
+.text-input {
+    font-size: 1.5vh;
+    max-height: 5vh;
+    background-color: #131F24;
+}
+
+.form-select {
+    min-width: 100%;
+    max-height: 5vh;
+    font-size: 2vh !important;
+    background-color: #131F24;
+    color: #FFF;
+    margin: 0;
+}
+
+.right-container {
+    position: fixed;
+    top: 0;
+    right: 2rem;
+    bottom: 2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    gap: 20px;
+  }
+
+.bottom-right-container {
+  background-color: #131F24;
+  color: #D1D1D1;
+  border-radius: 12px;
+  padding: 2rem;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  width: 20vw;
+  max-width: 300px;
+  max-height: 55vh;
+  font-size: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: auto; /* Ustawienie na dole kontenera */
+  border: 1px solid #2F7A7E;
+}
