+++ frontend-argonout/src/components/map/basicMap.css
@@ -22,6 +22,7 @@
 
 .popupContainer h5 {
   margin-bottom: 10px;
+  font-weight:bold;
 }
 
 .popupContainer img {
@@ -37,10 +38,10 @@
 .popupContainer a {
   text-decoration: none;
   font-weight: bolder;
-  color : #276304 !important;
+  color : #008caf !important;
   /* color : #29005e !important; */
 }
 .popupContainer a:hover {
-  color: #008caf !important;
+  color: #3ac0c7 !important;
   /* color: #5e1db3 !important; */
 }
